+++ CHANGELOG.rst
@@ -4,6 +4,7 @@ CHANGELOG
 UNRELEASED
 ----------
 
+* 🎉 feat(strings-truncate): Truncate function added
 * 🐛 fix(internal-utilities): isCustomSeparator function added to internal package
 * ✅ test(strings-words_test): Words testcases and benchmarks added
 * 🎉 feat(strings-words): Words function added